requirements.txt
altair==5.5.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
crypto==1.4.1
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
Naked==0.1.32
narwhals==2.11.0
numpy==2.3.4
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.0.0
plotly==6.4.0
protobuf==6.33.0
psycopg2-binary>=2.9.9
pyaes==1.6.1
pyarrow==21.0.0
pycryptodome==3.23.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
schedule==1.2.2
shellescape==3.8.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
streamlit==1.51.0
streamlit-autorefresh==1.0.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tuya-connector-python==0.1.2
tuya-iot-py-sdk==0.6.6
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
websocket-client==1.9.0
